requirements.txt
azure-identity 
requests
msfabricpysdkcore

ms-fabric-cli